first release

+++ packer.py
@@ -1,50 +1,68 @@
-#!/bin/python
+#!/usr/bin/env python
 import sys, os
 
 if len(sys.argv) != 3:
-	print "usage: %s exefile tarpath" % sys.argv[0]
+	print "Usage: %s elfname tarname" % sys.argv[0]
+	print "\telfname: the elf binary file to patch"
+	print "\ttarname: the final package name"
+	print "\teg: ./packer.py /bin/ls pack"
 	sys.exit(-1)
 
-exefile = sys.argv[1]
-tarpath = sys.argv[2]
+elfname = sys.argv[1]
+tarname = sys.argv[2]
 
-patchelf = "patchelf"
-basepath = "./%s" % tarpath
-command = r'ldd %s | sort | grep -oE "=> /[^(]*" | grep -oE "/[^(]*"' % exefile
-linkerpath = "linker"
+tmppath = "TEMP_PATH"
+basepath = "./%s/%s" % (tmppath, tarname)
+
+patchelf = "./patchelf/patchelf"
+command = r'ldd %s | sort | grep -oE "=> /[^(]*" | grep -oE "/[^(]*"' % elfname
+loaderpath = "loader"
 libpath = "libs"
 
-print "mkdir"
-os.system("rm -rf %s" % basepath)
-os.makedirs("%s/%s" % (basepath, linkerpath))
-os.makedirs("%s/%s" % (basepath, libpath))
-path = "%s/%s" % (basepath, libpath)
+print "test"
+if !os.path.isfile(patchelf):
+	print "You should compile patchelf first!"
+	print "You can read README for more information."
+	sys.exit(-1)
 
+print "mkdir"
+os.system("rm -rf %s" % tmppath)
+os.makedirs("%s/%s" % (basepath, loaderpath))
+os.makedirs("%s/%s" % (basepath, libpath))
+
+print "cp libs"
+path = "%s/%s" % (basepath, libpath)
 for lib in os.popen(command).readlines():
 	cpcommand = "cp -L %s %s" % (lib.strip(), path)
 	os.system(cpcommand)
 	print cpcommand
 
-print "cp exefileram"
-os.system("cp -L %s %s" % (exefile, basepath))
-newexefile = "%s%s" % (basepath, exefile[exefile.rindex('/'):])
+print "cp elffile"
+cpcommand = "cp -L %s %s" % (elfname, basepath)
+os.system(cpcommand)
+print cpcommand
 
 print "modify rpath"
-os.system(r"%s --force-rpath --set-rpath \$ORIGIN/%s %s" % (patchelf, libpath, newexefile))
+elfonly = elfname[elfname.rindex('/'):]
+os.system(r"%s --force-rpath --set-rpath \$ORIGIN/%s %s%s" % (patchelf, libpath, basepath, elfonly))
 
-linker = os.popen("%s --print-interpreter %s" % (patchelf, exefile)).read().strip()
-os.system("cp -L %s %s/%s" % (linker, basepath, linkerpath))
-print "linker:", linker
+print "cp loader"
+loader = os.popen("%s --print-interpreter %s" % (patchelf, elfname)).read().strip()
+os.system("cp -L %s %s/%s" % (loader, basepath, loaderpath))
+print "loader:", loader
 
-os.chdir(basepath)
-print "modify linker"
-newlinker = "./%s%s" % (linkerpath, linker[linker.rindex('/'):])
-os.system(r"%s --set-interpreter %s .%s" % (patchelf, newlinker, newexefile))
+print "create script"
+shname = "%s/%s.sh" % (basepath, elfonly)
+os.system(r"echo '#!/bin/sh' > %s" % shname)
+os.system(r"echo './%s%s .%s $@' >> %s" % (loaderpath, loader[loader.rindex('/'):], elfonly, shname))
+os.system("chmod +x %s" % shname)
 
-os.chdir("../")
+os.chdir(tmppath)
 print "create tar"
-os.system("tar czf %s.tar.gz %s" % (tarpath, tarpath))
-os.system("rm -rf %s" % tarpath)
+os.system("tar czf %s.tar.gz %s" % (tarname, tarname))
+os.system("mv %s.tar.gz ../" % tarname)
+os.chdir("../")
+os.system("rm -rf %s" % tmppath)
 
 print "done"
